Add files via upload
To get this to work, download xampp. Once xampp is installed, go to the xampp folder in your C: directory. Inside it there is a htdocs folder. Thats basically a public_html for xampp. Inside it, drop the whole gp folder in there. Go to the mail xampp directory, scroll to the bottom and open xampp-control. When it pops up, click start for Apache and for MySQL. That should start the php server. Then open your web browser and type "localhost/phpmysamin". That will open phpmyadmin to see the sql. Click the SQL tab at the top and paste the contents of the ooplr.sql file into the SQL query and click go at the bottom right. This will make a users database with 3 tables in it for the user accounts. After that, it should all work, "localhost/gp/index.php" should get you to the main page.